internal/blocktest: add package for shared test code (#27270)
